asr/requirements.txt
moviepy==1.0.3
gentle
modelscope
huggingface_hub
flask
funasr
numpy<2.0.0
spacy
openai-whisper
requests
# for spaCy installing, refer to https://spacy.io/usage
